+++ Makefile
@@ -4,10 +4,10 @@ venv:
 reqs: venv
 	./venv/bin/pip install -r requirements.txt && touch reqs
 
-jira_agile_extractor.egg-info: reqs
+agile_analytics.egg-info: reqs
 	./venv/bin/python setup.py develop
 
-test: jira_agile_extractor.egg-info
+test: agile_analytics.egg-info
 	./venv/bin/py.test -svv --flake8
 	# Poor man's system test, not committed because it requies a real jira
 	[ -f tryout.py ] && ./venv/bin/python tryout.py
@@ -17,7 +17,7 @@ clean_pycs:
 
 clean: clean_pycs
 	rm -rf venv
-	rm -rf jira_agile_extractor.egg-info
+	rm -rf agile_analytics.egg-info
 
 docs: reqs
 	cd docs && make html
