+++ tests/test_analyzers.py
@@ -6,5 +6,5 @@ import pytest
 @pytest.mark.fixture
 def klass():
     """Return the Class Under Test."""
-    from jira_agile_extractor.analyzers import ThroughputAnalyzer
+    from agile_analytics.analyzers import ThroughputAnalyzer
     return ThroughputAnalyzer
